+++ frontend/src/metabase/dashboard/hooks/use-dashboard-url-params.ts
@@ -1,4 +1,5 @@
 import type { Location } from "history";
+import { useEffect } from "react";
 
 import {
   useDashboardFullscreen,
@@ -7,7 +8,9 @@ import {
 } from "metabase/dashboard/hooks";
 import { useLocationSync } from "metabase/dashboard/hooks/use-location-sync";
 import type { RefreshPeriod } from "metabase/dashboard/types";
-import { useEmbedFrameOptions, useSetEmbedFont } from "metabase/public/hooks";
+import type { DashboardUrlHashOptions } from "metabase/dashboard/types/hash-options";
+import { parseHashOptions } from "metabase/lib/browser";
+import { useEmbedFont, useEmbedFrameOptions } from "metabase/public/hooks";
 import type { DisplayTheme } from "metabase/public/lib/types";
 
 import { useAutoScrollToDashcard } from "./use-auto-scroll-to-dashcard";
@@ -19,7 +22,7 @@ export const useDashboardUrlParams = ({
   location: Location;
   onRefresh: () => Promise<void>;
 }) => {
-  useSetEmbedFont({ location });
+  const { font, setFont } = useEmbedFont();
 
   const {
     background,
@@ -65,6 +68,12 @@ export const useDashboardUrlParams = ({
     location,
   });
 
+  useEffect(() => {
+    const { font } = parseHashOptions(location.hash) as DashboardUrlHashOptions;
+
+    setFont(font ?? null);
+  }, [location.hash, setFont]);
+
   return {
     isFullscreen,
     onFullscreenChange,
@@ -77,6 +86,8 @@ export const useDashboardUrlParams = ({
     background,
     bordered,
     titled,
+    font,
+    setFont,
     theme,
     setTheme,
     hideParameters: hide_parameters,
