+++ frontend/src/metabase/public/containers/PublicOrEmbeddedDashboard/PublicOrEmbeddedDashboardPage/PublicOrEmbeddedDashboardPage.tsx
@@ -41,6 +41,7 @@ export const PublicOrEmbeddedDashboardPage = (props: WithRouterProps) => {
     onRefreshPeriodChange,
     theme,
     titled,
+    font,
     locale,
   } = useDashboardUrlParams({ location, onRefresh: refreshDashboard });
 
@@ -67,6 +68,7 @@ export const PublicOrEmbeddedDashboardPage = (props: WithRouterProps) => {
         downloadsEnabled={downloadsEnabled}
         theme={theme}
         titled={titled}
+        font={font}
         parameterQueryParams={parameterQueryParams}
         cardTitled={true}
         locale={canWhitelabel ? locale : undefined}
