+++ frontend/src/metabase/public/hooks/index.ts
@@ -1,2 +1,2 @@
 export * from "./use-embed-frame-options";
-export * from "./use-set-embed-font";
+export * from "./use-embed-font";
